+++ truck-meshalgo/tests/filters/optimizing.rs
@@ -165,7 +165,7 @@ fn put_together_same_attrs_test() {
         })
         .collect();
 
-    mesh.put_together_same_attrs();
+    mesh.put_together_same_attrs(TOLERANCE);
     mesh.faces()
         .face_iter()
         .flatten()