+++ CHANGELOG.md
@@ -4,6 +4,7 @@ The version is of the bottom crate `truck-rendimpl`.
 
 ## Unreleased
 
+- Tolerance setting of `put_together_same_attrs`.
 - Increase the capacity of step file in adhoc-viewer.
 - Fix clippy.
 - Add tessellate test with ABC Dataset.